moving clibrary to new file
